Added Day 38
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible (Todays flood uses a proxy due to all websites being put in russian access only)
Mesage posted in the Ukraine IT Army for Day 38:
We see that some goals are beginning to come to life. Therefore, we offer an adjusted list of goals to focus our efforts as much as possible. And of course do not forget about the VPN!