Added Day 40 p2 (New Targets)
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible (Todays flood uses a proxy due to all websites being put in russian access only)
Mesage posted in the Ukraine IT Army for part 2 of Day 40:
We have to admit that after our previous attack Atol has significantly improved its security procedures. Now it is a really tough nut to crack but we made them spend extra money on that!
We are switching to the new targets, namely to the marketplace platforms and advertising websites.