Added Day 42
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible
Mesage posted in the Ukraine IT Army for Day 42:
Keep up the good work!
Due to our well-synchronized activity, the marking system has been unavailable for most companies since yesterday morning. The aggressor country's Ministry is trying to manage the process for large corporations manually and recommends stocking up with the codes to last through the week. But they won't be able to manage hundreds of thousands of companies that way.
Please try to use the aggressor country's VPN.