Added Day 43
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 43:
The attack was highly productive. The marking system was so malfunctioning that the aggressor country's Ministry of Industry and Trade decided not to mark certain goods and allow distribution of the unmarked goods up until April 11.
Thank you all for joining the attack.
Today we'll focus on online TV broadcasting services to ensure they are knocked off s well.