Added Day 47
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 47:
We did a great job yesterday! The marking system still hasn't recovered.
Thanks a lot for your help!
As a result of the attack, the system was blocked. And the Ministry of Trade issued a recommendation to keep a one-week stock of markings. Thus they just passed the buck to the business (who were not happy about it).
Let's keep the system down for one more day!