Added Day 48
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 45:
We've been showing great results with the marking system for two days in a row! The majority of its customers can use it.
We'll pause our activities related to this system for now. But we'll keep monitoring it to renew them when needed.
Thank you for your help!
Today we'll remind the banking system about ourselves. The banking industry is a very tough nut to crack, especially after multiple attacks from our end. We are much more powerful now. So let's hit the targets below, don't forget that you can use the "Death by 1000 needles" solution and that VPN is an absolute must!