Added Day 49
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 49:
Today we have an awesome target that we haven't attacked before. It is a platform for collecting fees for using the aggressor country's road system by trucks heavier than 12 tons.
Let's do our best to block those transactions and thus block a source of income for the state budget.