Added Day 52
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 52:
It's the beginning of a workweek, and we switching back to the system of veterinary inspections.
We've targeted them before so it will be much harder this time.
We need to use VPNs and proxy servers (ideally the aggressor country's ones). 
Let's throw ourselves on the following targets with all our determination.