Added New IP's for Day 52
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
(Info box showing your public IP is now displayed a top center of screen.)
Mesage posted in the Ukraine IT Army for Day 52 (New IP's):
We see considerable efforts to protect the marking system. They even published an interview with their tech specialist. So let's pay a visit to the marking system while still not putting the veterinary system aside. Here is the list of current targets.