Added Day 53
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 53:
Today we'll focus on aviation. We've already attacked some of the targets before, and thus they are more ready for our comeback. But we know how to knock out their defences. We need to unite our efforts and push through with all we have.
After all, why do they need to travel by plane when they can do it on foot. 
Don't forget to use a VPN for your attacks (especially for this one). If possible, choose Belarus in its settings.