Added Day 55
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 55:
You did a great job. Some of the fiscal data operators are down; others have hidden behind the geo-blocking. So we'd ask you to push forward and try to increase efficiency by using VPNs or proxy servers from Russia, Belarus and Kazakhstan.
In addition to it, we have a new and critical target - external resources for auditing and accounting service, "1C-Reports". Blocking them would be a significant impediment to timely tax reporting.