Added Day 56
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 56:
Well done! There is an increasing number of complaints about the "1C-Reports" service. We also wanted to remind you that it's crucial to use Russian VPN or proxy servers and that you can find the attacks guidelines and other helpful information on our website: https://itarmy.com.ua/instruction/?lang=en
To consolidate the progress we've made, we'll keep on targeting the critical resources of the service (see an updated list below).