Added Day 59
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 59:
The aggressor country has extended the deadline for submitting taxes until April 26th. But we'll make sure that they are still unable to file them on time.
We've updated the target list and added the "Contour" system to it since this system is also used for tax recording. If you use "Death by 1000 needles", the updated configuration will include the most up-to-date target list.
We'd like to remind you that attacks guidelines and other information can be found on our website: https://itarmy.com.ua/instruction/?lang=en