Added Day 60
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 60:
Great job everyone! Judging by the official appeal of 1C to their customers, they can not normally cope with our attacks for as long as 5 days! This is also confirmed by the many comments of their very angry users :) 
Since today is the final day of reporting, we need to spend it as productively as possible! And tomorrow we promise a new and interesting goal ;) 
We take the resources from previous publications and we keep attacking them : 
https://t.me/itarmyofukraine2022/309 https://t.me/itarmyofukraine2022/311
We remind you that you can find instructions on attacks and other information on our website: https://itarmy.com.ua/