Added Day 66
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 65:
Today's target was received from one of our subscribers. "EGAIS" is a Russian federal state-owned automatic system for tracking the production and distribution of alcoholic beverages in the country. If this system is down, the official turnover of liquor and spirits in Russia will be blocked!
Just a quick reminder that attacks guidelines and other information can be found on our website: https://itarmy.com.ua/instruction/?lang=en