Added Day 67
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 67:
Based on the services' availability examination and the comments from their users, we are on the right track. Today we suggest consolidating (or even advancing) our progress by using VPN/proxy servers specifically from Russia.