Added Day 69
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 69:
Russian media reports about our progress! They've also mentioned that protracted disruptions in "EGAIS" service would lead to suspending operations and thus result in decreased income to the aggressor country's budget. That sounds like a good reason to push on!