Added Day 70 and Day 71 as clones of Day 69
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 70 and 71 (clones of Day 69 due to a mix up with the days):
Based on the services' availability examination and the comments from their users, we are on the right track. Today we suggest consolidating (or even advancing) our progress by using VPN/proxy servers specifically from Russia.
Russian media reports about our progress! They've also mentioned that protracted disruptions in "EGAIS" service would lead to suspending operations and thus result in decreased income to the aggressor country's budget. That sounds like a good reason to push on!