Added Day 72
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 72:
Note: the list of "EGAIS" resources has been changed because Rushists attempted to shield themselves from the attacks! A critical part of services is hidden behind the geoblock, so we need to use VPNs and proxies from Russia only!
And in addition to "EGAIS," let's put down bidding platforms.