Added Day 74 and removed sideways scroll
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 74:
We have a tradition to combat Russian propaganda on Sunday. Let's stick to it. Additionally, we'd like to keep a firm grip on "EGAIS" because due to your help, there is an ongoing flow of complaints about this service.