Added Day 75 + EGAIS
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 75:
Let's switch to targets that are related to today's holiday in Russia.
And don't forget to keep "EGAIS": https://t.me/itarmyofukraine2022/321
EGAIS:
"EGAIS" is a Russian federal state-owned automatic system for tracking the production and distribution of alcoholic beverages in the country. If this system is down, the official turnover of liquor and spirits in Russia will be blocked!