Added day 76
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 76:
Today we'll continue our fight against propaganda. In addition to it, we'll "pay some attention" to the list of shoe manufacturers (that was recently shared with us) that supply the Russian army with boots.