Added Day 84
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 84:
Today we'll work on the recently launched app store "NashStore" (OurStore).

In addition to it, let's consolidate our previous achievements by keeping the resources related to airlines, local media and digital signatures down. The list of the targets was updated and added to both "Death by 1000 needles" and "mhddos_proxy".

By the way, there was a rollout of the major updates to "mhddos_proxy" yesterday. It's now switched to an asynchronous version (that doubled the number of threads via -t) and increased traffic.