Added Day 86
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 86:
Today we'll start working on russian online banking. Below is the first batch of targets for a warm-up.
There was a rollout of the major updates to mhddos_proxy. It's now switched to an asynchronous version (that doubled the number of threads via -t) and added --copies setting to run multiple copies.