Added Day 87
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 87:
Let's continue our attacks on the russian online banking. The list of targets for today is much longer. Russian banks are not limited to Sberbank, Alfa or VTB. Smaller banks also finance the war against Ukraine and thus deserve our attention!