Added Day 88
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 88:
We considered your suggestions and added several dozens of Russian banks to the new list of targets. You were concerned that the attacks on banks were scheduled for a weekend. But that timing allows us to keep them down for much longer since their tech support does not work over weekends. Client satisfaction and reputation are not the priorities in Russia.
We also wanted to remind you about important updates to mhddos_proxy that increased the efficiency of our attacks. Below is a link to the instruction for the installation of updates. And don't forget that you can grant access to your cloud resources to our bot that will synchronize the attacks launching. You can read more about it on our website.