Added Day 90
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Mesage posted in the Ukraine IT Army for Day 90:
Let's consolidate the achievements made over the last couple of days and keep the Russian banking and microfinance industry down. Below is a list of target updates.
Here is a link (https://gist.githubusercontent.com/ddosukraine2022/f739250dba308a7a2215617b17114be9/raw/all_targets.txt) to the updated list of targets. If you use "Death by a 1000 needles" or mhddos_proxy, their updated configurations will include the most up-to-date targets out-of-the-box. And don't forget to tag --itarmy when launching the mhddos_proxy. By the way, mhddos_proxy has English localization now; you can switch to it by using the --lang en settings.