Added Day 92
Added Day 92Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 92:
Yesterday we were busy with the banking industry and didn't pay a visit to RuStore by VK (it's yet another copy of Google Play). So let's catch up on it today!
Banks and microfinance institutions under attack remain a valid target.