Added Day 93
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 93:
Russian app store RuStore failed to handle our 'greetings' and has applied geoblocking that affected Russian VPN servers as well. Let it be.
Instead, we'll add Russian insurance companies to our list of targets.