Added Day 95
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 95:
We have good results, so today we have added a second batch of russian propaganda targets. Here is a link to the updated list of targets.