Added Day 96
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 96:
For the next couple of days, we'll attack Russian internet providers and consolidate the progress made on previous targets. Have you paid your internet service bills? :)
Here is a link to the updated list of targets.