Added Day 99
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 99:
We have already successfully attacked the russian information system: Chesnyi Znak, Platon, and EGAIS were paralyzed for a long time, that contributed to the 'negative growth' of the Russian economy. Today, we publish a list of similar systems that will be our targets for the next couple of days.