Added Day 100
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 100:
You have shown good results on yesterday's goals! One part lies, and the other is hidden behind a geoblock. Today we are connecting to the attacks of another information system of the Russian Federation.