Added Day 103 - Fixed double colon on one of the targets
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 103:
On the night of June 5th, hackers breached the Russian ministry of construction, housing and utilities infrastructure website. They placed a
"Glory to Ukraine" sign in the header and warned that they would make the ministry's database publicly available if they were not paid a ransom of 0,5 bitcoin to reward their efforts. Let's join the attacks on the russian public resources. They seem to get way too relaxed :)