Added day 106
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 106:
Our new target for today is an online part of the "Moi Documenty" (My Documents) service that allows russian citizens to use public or municipal services. Most of the previous resources are down and are still included in our list of targets.