Added Day 108
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 108:
As we promised, we are adding new russian media to our list of resources. Those media have already experienced some problems (we have contributed to them) with quick coverage of targets we chose last week. BTW, recently the russian foreign ministry called us the most active group.