Added Day 109
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 109:
Also today let's support our colleagues attack on the following russian resources (tough nut to crack):
We continue to attack the russian media, government web resources, online platforms. We also remind the occupiers that Crimea is Ukraine!