Added Day 113
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 113:
We add the resources of arbitration courts to the list of targets. Also do not forget that the St. Petersburg "international" economic forum runs until June 18, so continue to "visit" it.