Added Day 127
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 127:
CRM needed by lots of Russian business which pay taxes then it should not be so easy in our humble opinion. We strongly believe Russian business should focus on stopping the war not profit.