Added Day 131
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 131:
No new targets for today, recycling previous targets: Well done! For the last two days, the russians have not been able to access their CRM. As always, on weekends we switch to russian mass media.