Added Day 138
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 138:
Thanks to you, the sites of russian cinemas rested for the weekend. Less tickets bought online mean less money in the budget for financing the war. Today, we are returning all power to the attack on the logistics of the russian federation.