Added Day 145
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 145:
Copter drone shops is the next valid target to diminish its availability for russian troops. Please let us know if you can share more web addresses with a special form on the website.