Added Day 148
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 148:
No new targets posted for this day, recycling yesterdays targets. In addition, we add russian military dealers and your recommendations for drone stores to the list of targets.