Added Day 150
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 150:
No new targets posted for this day, recycling yesterdays targets. New Message was however posted..:
We continue to block the possibility of online purchase of missing equipment for russian orks. We are also attacking the russian ruling political party "united russia", known as putin's party, which is currently conducting propaganda activities in the temporarily occupied territories.