Added Day 153
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 153:
How about hitting the websites of artemiy lebedev, the propagandist, and marauder used to call a designer? He "achieved" being listed by Myrotvorets ("Piecemaker," the website features war criminals) and sanctioned by the National Security and Defense Council of Ukraine even before the great invasion this year.