Added Day 162
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 162:
Now new list of targets today, recycling yesterdays list: Today's target is the pension fund of the russian federation. Why? Because through it, russian military families receive payments and compensation.