Added Day 163
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 163
Now new list of targets today, recycling previous list: Today's target is the pension fund of the russian federation. Why? Because through it, russian military families receive payments and compensation.