Added Day 167
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new message was posted in the Ukraine IT Army for Day 167 recycling previous targets:
We challenge import phaseout players of the aggressor. TrueConf is the platform for video calls and remote works. It is quite popular among private and public companies in russia recently.