Added Day 169
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new targets or message was posted in the Ukraine IT Army for Day 169 as of yet, recycling previous targets (will update if new list is released):
We challenge import phaseout players of the aggressor. TrueConf is the platform for video calls and remote works. It is quite popular among private and public companies in russia recently.