Added Day 170
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 170
Message posted with new set of Targets in telegram by the IT Army of Ukraine:
Today, other russian video conferencing products are under attack. Let the weekend come early for their customers.