Added Day 182
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 182:
The news that on the russian job search site SuperJob four regions of Ukraine, which are currently temporarily occupied by orcs, are marked as the territory of the russian federation did not pass by us. Today they are our target. We see they are hiding behind a lot of protections. This won’t help them! Working together we can accomplish anything!