Added Day 183
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 183:
SuperJob responded with errors instead of occupied areas of Ukraine after our attacks yesterday. 
We have not touched the top propaganda media as TASS. Do you think domestic ddos-protection can be of help?