Added Day 185
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
Message posted in the Ukraine IT Army for Day 185
No new list today (yet), recycling previous targets with new set of Targets in telegram by the IT Army of Ukraine, will update if new IP list is released:
SuperJob responded with errors instead of occupied areas of Ukraine after our attacks yesterday. 
We have not touched the top propaganda media as TASS. Do you think domestic ddos-protection can be of help?