Added Day 216
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New Message posted in the Ukraine IT Army for Day 216, IP's re-used from day91
We will open a trading session on the Moscow Exchange with a DDoS attack on them. A report on yesterday's attack on gas station systems is coming soon.