Added Day 217
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
new Message posted in the Ukraine IT Army for Day 217, however target URL's not updated, recycling previous targets:
Yesterday we paralyzed the work of personal accounts of gas stations and fuel cards of russians. The attack turned out to be surprisingly successful and sensitive, so we decided to switch to them again today after the morning "trading" on the moscow stock exchange.