Added Day 219
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
new Message posted in the Ukraine IT Army for Day 219:
Russian banks and payment systems are constantly under our pressure. The next ones are Gosznak for securities productions, and Rosinkas the biggest incassation service, GIIS DMDK the diamonds and precious metals labeling authority.