Added Day 220
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 220:
You could see a huge number of videos where newly minted russian mobs (mobilized russian citizens) are obliged to provide themselves with everything they need. For them, the easiest option is to google the nearest or with delivery voentorg (war merchant), so our job is to keep them offline and do not allow to equip the enemies with the necessary equipment.