Add files via upload
**Re-Uploaded to updated target notes**
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 222:
In addition to specialized internet shops, drones are also sold by large networks, so they are also under our sights! Today we are attacking citilink, which has more than 1,000 stores and points of delivery of orders throughout russia. Unavailability of this online store is also a blow to the economy of the russian federation.