Added Day 224
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 224:
We continue the theme of negative provision and unite the attack on military stores and specialized online drone stores. Today, the survivors and potentially vulnerable are targeted, with a total of 72 stores. Others are unavailable or have geo blocked themselves from the civilized world. Our attacks on military and drone stores are also reported in the russian media, which confirms the feasibility of attacking them.