Added Day 226
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 226:
Sberbank has a service for entering websites and applications - Sber ID. We can paralyze the entire eco-system of the bank by knocking it out of action. Get down to business!