Added Day 227
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 227:
Sberbank reported on yesterday's problems with clients. They say there were technical works... But it is very strange that they coincided with our attack.
We will also dedicate today to "happy" customers in the russian federation and take "wildberries" marketplace to work.