Added Day 228
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new targets posted as of yet today, recycling previous targets
New message posted in the Ukraine IT Army for Day 228:
Even the geoblock knows that Crimea is Ukraine! Wildberries did not come up with anything better than to add a geoblock, but this also plays into our hands because they will not work in the occupied territories of Ukraine.