Added Day 230
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 230:
Today we continue the topic of banks. We share screenshots with the results that are displayed in russian media and resource status check service.