Added Day 231
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 231:
Let's return to the topic of exchanges. It will not be possible to come to the stock exchange and trade on your own, you need a broker who will act as an intermediary between the investor and the company whose securities are sold. So today the finam.ru and solidbroker.ru brokers and the banki.ru financial marketplace, which has lists and ratings of brokers in russia, are under scrutiny.