Added Day 234
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new message posted in the Ukraine IT Army for Day 234, recycling yesterdays targets and message:
Today we return to banks again. Tinkoff Bank in the sights.
We aim to interrupt payment processing, delay liabilities, undermine bank’s reputation as usual.