Target update for Day 234
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 234:
We return to the traditional fight against propaganda on the weekend. You know how poisonous TV is in Russia. So we will hold an evening prevention program for the viewers of Ivi.