Added Day 235
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new message posted in the Ukraine IT Army for Day 235, recycling yesterday's targets & message:
We return to the traditional fight against propaganda on the weekend. You know how poisonous TV is in Russia. So we will hold an evening prevention program for the viewers of Ivi.