Added Day 261 and updated List for Days (bulk TL)
Update for days and list of Days, List for Days Updated (TL)
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message was posted in the Ukraine IT Army for Day 261:
On the day of shopping, we will reduce the planned sales on russian trading platforms and large online stores. Let's start with Ozon.