Updated Day 268 targets
Update targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new message was posted in the Ukraine IT Army for Day 267, recycling previous targets:
We have previously carried out attacks on military and drone stores. Today, we are supplementing this topic with stores of radio-electronic parts that are in demand among the military of the russian federation. We start with a large store - "CHIP and DIP"