Updated Day 271 - Targets & days (TL)
Updated targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New message posted in the Ukraine IT Army for Day 271:
Last week, we worked productively with 1C services. Today we'll start working on bidding platforms. RTS-tender in the sights.