Updated Day 280 - Targets & days (TL) Eng + Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible.
No new message posted in the Ukraine IT Army for Day 280, recycling yesterday's targets:
Alfa Bank learns to lie from the leadership of its state and deletes angry comments from customers for the sake of a positive atmosphere in the group.
The baton passes to VTB Bank. We aim to interrupt payment processing, delay liabilities, undermine bank’s reputation as usual.