Added Day 410 & 411 - Targets and Days (TL) Eng+Ukr
Updated targets for days, simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible.
Message Posted in the IT Army Telegram server for current targets:
Renaissance Insurance and Ingosstrakh have joined forces with SOGAZ the club "Find Our Services If You Can" and are amiably conducting technical work. Insurance, as is well known, derives from the word "fear" in Russian. So, is it worth insuring against the fear of war? Demand for this is expected to grow in the coming days, although vodka and the balalaika will be more reliable than these Western capitalist inventions of yours.
\**\ 
The Russian Alfastrakh insurance company is not working, claiming that the internet has run out. It's the karmic fate of the aggressor's insurance business. It's a pity that there's no electricity, but considering the direction of their country's movement, it's only a matter of time. Well done, ITtackers, great shot!
/ *** / 
"Ренесанс Страхування" та “Інгосстрах” склали компанію СОГАЗ клубі "Знайди наші сервіси, якщо зможеш" і дружньо проводять технічні роботи. Страхування, як відомо, походить від слова "страх". Тож, може варто страхувати від страху війни? Попит на це має зростати наступними днями, хоча горілка і балалайка будуть надійнішими, ніж ці ваші західні капіталістичні винаходи.
\**\Росіянський Альфастрах не працює, каже інтернет закінчився. Кармічна доля страхового бізнесу агресора. Шкода що не світло, але враховуючи вектор руху їх країни, то це тільки питання часу. Молодці айтірелисти, чудове влучання!