Updated Targets for Day 474 Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
No new targets posted, using targets from db1000n's github:
No targets have been publicly posted on the IT Army of Ukraine telegram for over a week now.
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
when/if the IT Army of Ukraine post new targets, we'll go from those with targets sourced our selves.