Updated Info box for day 531 | Оновлено Інформаційне поле для дня 531
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
According to The Times, President Biden's mornings start not with coffee, but with planning tasks for the IT ARMY. From the Tatarstan public service website, to the Russian Railways and the milk licensing system - nothing leaves him at peace. We're eagerly awaiting: what task has the president prepared for us today?
Meet the new utility ADSS from IT ARMY for Linux. Easy installation, integration with DDoS programs, quick start. Give it a try today!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Повідомлення оприлюднило IT Army of Ukraine:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub. Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут. Повідомлення оприлюднило IT Army of Ukraine:
Згідно статті в The Times, кожен ранок президента Байдена починається не з кави, а з планування завдань для IT ARMY. Від сайту держпослуг Татарстану, до РЖД і системи ліцензування молока - ніщо не дає йому спокою. З нетерпінням чекаємо: яке ж завдання пан президент приготував нам сьогодні?
Познайомтеся з новою утилітою ADSS від IT ARMY для Linux. Легкість встановлення, інтеграція з DDoS програмами, швидкий старт. Спробуйте вже сьогодні!