Added Day 540 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
❗❗❗Tsargrad channel (the one we recently shut down) reports: Ruzzian hacker group 'Solntsepek' hacked all of Ukraine's EW systems at once(!!!). So, this thing EW is, according to them, connected to the internet or maybe linked to some central server. No, no - there was a saboteur with a flash drive sneaking around.
P.S. They also found documentation on the construction of Columbia-class nuclear submarines in the Defense Intelligence of Ukraine. Kudos to our spies on this one, of course. 😂😂😂 Too bad they didn't find any photos of aliens.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Щоб не відставати від цілей, ми збираємося одночасно використовувати цілі db1000n із їхнього сховища GitHub. Ці оновлення відбуваються щодня, тож якщо db1000n зміниться, ймовірно, пройде ще кілька років, перш ніж зміни відобразяться тут. Повідомлення оприлюднило ІТ-Армія України:
❗❗❗Канал Царьград (той що ми нещодавно відключали) повідомляє: російська група хакерів "Солнцепьок" зламала одночасно(!!!) всі системи РЕБ в Україні! Тобто РЕБ на їх думку, підключено до інтернету чи може під'єднано до якогось центрального сервера. Ні, ні - там диверсант з флешкой прокрався.

P.S. Вони ще знайшли документацію з будівництва атомних підлодок класу Columbia в ГУР. Тут нашим розвідникам звісно респект. 😂😂😂 Шкода що фоток прибульців не знайшлось.