Added Day 544 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Integration of technologies into Ukraine's defense strategy is significantly impacting the course of the war. According to Palantir, its artificial intelligence is already aiding in target identification and making a substantial contribution on the battlefield. The extensive use of drones, AI tagging, and facial recognition systems are not just new technologies, but factors actively contributing to Ukraine's resilience and effectiveness in combat - writes Financial Times.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Щоб не відставати від цілей, ми збираємося одночасно використовувати цілі db1000n із їхнього сховища GitHub. Ці оновлення відбуваються щодня, тож якщо db1000n зміниться, ймовірно, пройде ще кілька років, перш ніж зміни відобразяться тут. Повідомлення оприлюднило ІТ-Армія України:
Інтеграція технологій у стратегію оборони України сильно впливає на хід війни. За твердженням компанії Palantir, її штучний інтелект вже допомагає у визначенні цілей і робить вагомий внесок на полі бою. Масове використання дронів, мітки з використанням ШІ та системи розпізнавання обличчя це не просто нові технології, а чинники які активно сприяють стійкості та ефективності України на полі бою - пише Financial Times.